refactor: split app.js into focused ES modules under public/src/
Mechanically split the 3,610-line public/app.js into 13 focused ES modules under public/src/, with public/app.js reduced to a one-line shim that imports ./src/main.js. This keeps the service worker's hardcoded /app.js cache entry valid while moving all logic into maintainable, single-responsibility files (state, definitions, ui, sdk-client, auth, contracts, form/*, operations, result-format, execute, main).

Function bodies are moved verbatim with no renames, inlining, or behavior changes; only export/import lines were added. Shared mutable state (state, elements, dynamicInputHandlers) is exported by reference so every module sees the same live objects. The callEvo dispatcher and formatResult move intact.

Verified byte-identical bodies (multiset line diff: 0 differences), a closed import graph (0 missing/unused imports), clean node --check on all files, and a passing test suite (smoke, query, transition).
